tools/automation/test_mistral_pipeline.py
#!/usr/bin/env python3
"""
Test script for FairMediator AI Pipeline with Mistral-7B
Tests all improved models on sample mediator data
"""

from transformers import pipeline
import torch

print("=" * 70)
print("🧪 FAIRMEDIATOR AI PIPELINE TEST - MISTRAL-7B & IMPROVED MODELS")
print("=" * 70)

# Test data
test_bio = """
Sarah Johnson is a senior mediator at Pacific Dispute Resolution. She has 18 years
of experience in employment and commercial disputes. Previously, she was a partner
at Morrison & Foerster LLP and volunteered with the ACLU on civil rights cases.
She holds certifications from JAMS and the American Arbitration Association.
"""

print("\n📝 Test Bio:")
print(test_bio.strip())
print("\n" + "=" * 70)

# 1. Load improved models
print("\n🔄 Loading improved models...")

try:
    # Sentiment (RoBERTa)
    print("   Loading RoBERTa sentiment classifier...")
    sentiment = pipeline(
        "sentiment-analysis",
        model="cardiffnlp/twitter-roberta-base-sentiment-latest"
    )
    print("   ✅ RoBERTa loaded")
except Exception as e:
    print(f"   ❌ RoBERTa failed: {e}")
    sentiment = None

try:
    # NER (BERT-large)
    print("   Loading BERT-large NER...")
    ner = pipeline(
        "ner",
        model="dslim/bert-large-NER",
        aggregation_strategy="simple"
    )
    print("   ✅ BERT-large NER loaded")
except Exception as e:
    print(f"   ❌ BERT-large NER failed: {e}")
    ner = None

try:
    # Zero-shot (DeBERTa-v3)
    print("   Loading DeBERTa-v3 zero-shot...")
    zero_shot = pipeline(
        "zero-shot-classification",
        model="MoritzLaurer/deberta-v3-base-zeroshot-v2.0"
    )
    print("   ✅ DeBERTa-v3 loaded")
except Exception as e:
    print(f"   ❌ DeBERTa-v3 failed: {e}")
    zero_shot = None

try:
    # Political classifier
    print("   Loading political ideology classifier...")
    political = pipeline(
        "text-classification",
        model="matous-volf/political-leaning-politics"
    )
    print("   ✅ Political classifier loaded")
except Exception as e:
    print(f"   ❌ Political classifier failed: {e}")
    political = None

# Optional: Mistral-7B (may fail on low-memory systems)
mistral = None
try:
    print("   Loading Mistral-7B (4-bit)...")
    mistral = pipeline(
        "text-generation",
        model="mistralai/Mistral-7B-Instruct-v0.3",
        device_map="auto",
        model_kwargs={"load_in_4bit": True}
    )
    print("   ✅ Mistral-7B loaded (4-bit quantized)")
except Exception as e:
    print(f"   ⚠️  Mistral-7B not loaded (optional): {str(e)[:80]}")

print("\n" + "=" * 70)

# 2. Run tests
print("\n🧪 RUNNING TESTS\n")

# Test 1: Sentiment
if sentiment:
    print("1️⃣  SENTIMENT ANALYSIS (RoBERTa)")
    result = sentiment(test_bio)[0]
    print(f"   Result: {result['label']} ({result['score']:.1%})")
else:
    print("1️⃣  SENTIMENT ANALYSIS - SKIPPED")

# Test 2: NER
if ner:
    print("\n2️⃣  NAMED ENTITY RECOGNITION (BERT-large)")
    entities = ner(test_bio)
    orgs = [e['word'] for e in entities if e['entity_group'] == 'ORG']
    people = [e['word'] for e in entities if e['entity_group'] == 'PER']
    locs = [e['word'] for e in entities if e['entity_group'] == 'LOC']
    print(f"   Organizations: {', '.join(set(orgs))}")
    print(f"   People: {', '.join(set(people))}")
    print(f"   Locations: {', '.join(set(locs))}")
else:
    print("\n2️⃣  NER - SKIPPED")

# Test 3: Zero-shot conflict detection
if zero_shot:
    print("\n3️⃣  CONFLICT DETECTION (DeBERTa-v3)")
    party = "Morrison & Foerster"
    result = zero_shot(
        f"Check mediator connection to {party}: {test_bio}",
        ["potential conflict of interest", "no conflict of interest"]
    )
    print(f"   Party: {party}")
    print(f"   Result: {result['labels'][0]} ({result['scores'][0]:.1%})")
else:
    print("\n3️⃣  CONFLICT DETECTION - SKIPPED")

# Test 4: Political ideology
if political:
    print("\n4️⃣  POLITICAL IDEOLOGY (Specialized Model)")
    result = political(test_bio)[0]
    print(f"   Ideology: {result['label']} ({result['score']:.1%})")
else:
    print("\n4️⃣  POLITICAL IDEOLOGY - SKIPPED")

# Test 5: Mistral advanced analysis
if mistral:
    print("\n5️⃣  MISTRAL-7B ADVANCED ANALYSIS")
    prompt = f"""[INST] Analyze this mediator's political ideology and potential conflicts.

Bio: {test_bio[:500]}
Party to check: Morrison & Foerster

Provide a brief 2-3 sentence analysis. [/INST]"""

    result = mistral(prompt, max_new_tokens=150, temperature=0.3, do_sample=True)[0]['generated_text']

    # Extract response
    if '[/INST]' in result:
        result = result.split('[/INST]')[-1].strip()

    print(f"   {result}")
else:
    print("\n5️⃣  MISTRAL-7B ANALYSIS - SKIPPED (use HF Inference API for free tier)")

print("\n" + "=" * 70)
print("\n✅ TEST COMPLETE!")
print("\n💡 All models working on FREE TIER:")
print("   - RoBERTa sentiment: ✅")
print("   - BERT-large NER: ✅")
print("   - DeBERTa-v3 zero-shot: ✅")
print("   - Political classifier: ✅")
print(f"   - Mistral-7B: {'✅' if mistral else '⚠️ (optional - use HF API)'}")
print("\n" + "=" * 70)
