tools/automation/requirements.txt
# Python Dependencies for FairMediator Automation
# ALL FREE - No paid services required!
# Install with: pip install -r requirements.txt

# Core Dependencies
python-dotenv>=1.0.1
requests>=2.32.0

# FREE Hugging Face Integration
huggingface-hub>=0.25.0
transformers>=4.47.0

# Web Scraping (FREE)
beautifulsoup4>=4.12.3
selenium>=4.27.0
playwright>=1.49.0

# Data Processing (FREE)
pandas>=2.2.3
numpy>=2.2.1

# MongoDB (FREE tier available)
pymongo>=4.10.1

# Scheduling (FREE)
schedule>=1.2.2

# Optional: Local inference (100% FREE, runs on your machine)
# torch>=2.5.1  # Uncomment if you want to run models locally
# sentencepiece>=0.2.0  # For tokenization
